+++ views/admin-page-header.php
@@ -35,21 +35,21 @@ $progress_planner_active_frequency = isset( $_GET['frequency'] ) ? \sanitize_tex
 		</button>
 		<?php
 		// Render the settings button.
-		\progress_planner()->get_popover()->the_popover( 'settings' )->render_button(
+		\progress_planner()->get_ui__popover()->the_popover( 'settings' )->render_button(
 			'',
 			\progress_planner()->get_asset( 'images/icon_settings.svg' ) . '<span class="screen-reader-text">' . \esc_html__( 'Settings', 'progress-planner' ) . '</span>'
 		);
 		// Render the settings popover.
-		\progress_planner()->get_popover()->the_popover( 'settings' )->render();
+		\progress_planner()->get_ui__popover()->the_popover( 'settings' )->render();
 
 		// Render the subscribe form button and popover if the license key is not set.
 		if ( 'no-license' === \get_option( 'progress_planner_license_key', 'no-license' ) ) {
-			\progress_planner()->get_popover()->the_popover( 'subscribe-form' )->render_button(
+			\progress_planner()->get_ui__popover()->the_popover( 'subscribe-form' )->render_button(
 				'',
 				\progress_planner()->get_asset( 'images/register_icon.svg' ) . '<span class="screen-reader-text">' . \esc_html__( 'Subscribe', 'progress-planner' ) . '</span>'
 			);
 			// Render the subscribe form popover.
-			\progress_planner()->get_popover()->the_popover( 'subscribe-form' )->render();
+			\progress_planner()->get_ui__popover()->the_popover( 'subscribe-form' )->render();
 		}
 		?>
 		<div class="prpl-header-select-range">
