+++ views/page-widgets/suggested-tasks.php
@@ -79,10 +79,10 @@ $prpl_badge  = \progress_planner()->get_badges()->get_badge( Monthly::get_badge_
 		?>
 	<?php endif; ?>
 	<?php
-	\progress_planner()->get_popover()->the_popover( 'monthly-badges' )->render_button(
+	\progress_planner()->get_ui__popover()->the_popover( 'monthly-badges' )->render_button(
 		'',
 		\esc_html__( 'Show all my badges!', 'progress-planner' )
 	);
-	\progress_planner()->get_popover()->the_popover( 'monthly-badges' )->render();
+	\progress_planner()->get_ui__popover()->the_popover( 'monthly-badges' )->render();
 	?>
 </div>
